refactor: Move custom field handling into separate module
