+++ netbox/resource_netbox_ipam_ip_addresses.go
@@ -11,6 +11,7 @@ import (
 	netboxclient "github.com/smutel/go-netbox/v3/netbox/client"
 	"github.com/smutel/go-netbox/v3/netbox/client/ipam"
 	"github.com/smutel/go-netbox/v3/netbox/models"
+	"github.com/smutel/terraform-provider-netbox/v4/netbox/internal/customfield"
 )
 
 func resourceNetboxIpamIPAddresses() *schema.Resource {
@@ -51,7 +52,7 @@ func resourceNetboxIpamIPAddresses() *schema.Resource {
 				Computed:    true,
 				Description: "The content type of this IP address (ipam module).",
 			},
-			"custom_field": &customFieldSchema,
+			"custom_field": &customfield.CustomFieldSchema,
 			"description": {
 				Type:         schema.TypeString,
 				Optional:     true,
@@ -176,7 +177,7 @@ func resourceNetboxIpamIPAddressesCreate(ctx context.Context, d *schema.Resource
 	}
 
 	resourceCustomFields := d.Get("custom_field").(*schema.Set).List()
-	customFields := convertCustomFieldsFromTerraformToAPI(nil, resourceCustomFields)
+	customFields := customfield.ConvertCustomFieldsFromTerraformToAPI(nil, resourceCustomFields)
 	description := d.Get("description").(string)
 	dnsName := d.Get("dns_name").(string)
 	natInsideID := int64(d.Get("nat_inside_id").(int))
@@ -270,7 +271,7 @@ func resourceNetboxIpamIPAddressesRead(ctx context.Context, d *schema.ResourceDa
 			}
 
 			resourceCustomFields := d.Get("custom_field").(*schema.Set).List()
-			customFields := updateCustomFieldsFromAPI(resourceCustomFields, resource.CustomFields)
+			customFields := customfield.UpdateCustomFieldsFromAPI(resourceCustomFields, resource.CustomFields)
 
 			if err = d.Set("custom_field", customFields); err != nil {
 				return diag.FromErr(err)
@@ -394,7 +395,7 @@ func resourceNetboxIpamIPAddressesUpdate(ctx context.Context, d *schema.Resource
 
 	if d.HasChange("custom_field") {
 		stateCustomFields, resourceCustomFields := d.GetChange("custom_field")
-		customFields := convertCustomFieldsFromTerraformToAPI(stateCustomFields.(*schema.Set).List(), resourceCustomFields.(*schema.Set).List())
+		customFields := customfield.ConvertCustomFieldsFromTerraformToAPI(stateCustomFields.(*schema.Set).List(), resourceCustomFields.(*schema.Set).List())
 		params.CustomFields = &customFields
 	}
