+++ netbox/resource_netbox_ipam_ip_range.go
@@ -10,6 +10,7 @@ import (
 	netboxclient "github.com/smutel/go-netbox/v3/netbox/client"
 	"github.com/smutel/go-netbox/v3/netbox/client/ipam"
 	"github.com/smutel/go-netbox/v3/netbox/models"
+	"github.com/smutel/terraform-provider-netbox/v4/netbox/internal/customfield"
 )
 
 func resourceNetboxIpamIPRange() *schema.Resource {
@@ -30,7 +31,7 @@ func resourceNetboxIpamIPRange() *schema.Resource {
 				Computed:    true,
 				Description: "The content type of this prefix (ipam module).",
 			},
-			"custom_field": &customFieldSchema,
+			"custom_field": &customfield.CustomFieldSchema,
 			"description": {
 				Type:         schema.TypeString,
 				Optional:     true,
@@ -109,7 +110,7 @@ func resourceNetboxIpamIPRangeCreate(ctx context.Context, d *schema.ResourceData
 	endAddress := d.Get("end_address").(string)
 
 	resourceCustomFields := d.Get("custom_field").(*schema.Set).List()
-	customFields := convertCustomFieldsFromTerraformToAPI(nil, resourceCustomFields)
+	customFields := customfield.ConvertCustomFieldsFromTerraformToAPI(nil, resourceCustomFields)
 	description := d.Get("description").(string)
 	roleID := int64(d.Get("role_id").(int))
 	status := d.Get("status").(string)
@@ -169,7 +170,7 @@ func resourceNetboxIpamIPRangeRead(ctx context.Context, d *schema.ResourceData,
 			}
 
 			resourceCustomFields := d.Get("custom_field").(*schema.Set).List()
-			customFields := updateCustomFieldsFromAPI(resourceCustomFields, resource.CustomFields)
+			customFields := customfield.UpdateCustomFieldsFromAPI(resourceCustomFields, resource.CustomFields)
 
 			if err = d.Set("custom_field", customFields); err != nil {
 				return diag.FromErr(err)
@@ -262,7 +263,7 @@ func resourceNetboxIpamIPRangeUpdate(ctx context.Context, d *schema.ResourceData
 
 	if d.HasChange("custom_field") {
 		stateCustomFields, resourceCustomFields := d.GetChange("custom_field")
-		customFields := convertCustomFieldsFromTerraformToAPI(stateCustomFields.(*schema.Set).List(), resourceCustomFields.(*schema.Set).List())
+		customFields := customfield.ConvertCustomFieldsFromTerraformToAPI(stateCustomFields.(*schema.Set).List(), resourceCustomFields.(*schema.Set).List())
 		params.CustomFields = &customFields
 	}
