+++ netbox/resource_netbox_tenancy_tenant.go
@@ -11,6 +11,7 @@ import (
 	netboxclient "github.com/smutel/go-netbox/v3/netbox/client"
 	"github.com/smutel/go-netbox/v3/netbox/client/tenancy"
 	"github.com/smutel/go-netbox/v3/netbox/models"
+	"github.com/smutel/terraform-provider-netbox/v4/netbox/internal/customfield"
 )
 
 func resourceNetboxTenancyTenant() *schema.Resource {
@@ -37,7 +38,7 @@ func resourceNetboxTenancyTenant() *schema.Resource {
 				Computed:    true,
 				Description: "The content type of this tenant (tenancy module).",
 			},
-			"custom_field": &customFieldSchema,
+			"custom_field": &customfield.CustomFieldSchema,
 			"description": {
 				Type:         schema.TypeString,
 				Optional:     true,
@@ -94,7 +95,7 @@ func resourceNetboxTenancyTenantCreate(ctx context.Context, d *schema.ResourceDa
 
 	comments := d.Get("comments").(string)
 	resourceCustomFields := d.Get("custom_field").(*schema.Set).List()
-	customFields := convertCustomFieldsFromTerraformToAPI(nil, resourceCustomFields)
+	customFields := customfield.ConvertCustomFieldsFromTerraformToAPI(nil, resourceCustomFields)
 	description := d.Get("description").(string)
 	groupID := int64(d.Get("tenant_group_id").(int))
 	name := d.Get("name").(string)
@@ -156,7 +157,7 @@ func resourceNetboxTenancyTenantRead(ctx context.Context, d *schema.ResourceData
 			}
 
 			resourceCustomFields := d.Get("custom_field").(*schema.Set).List()
-			customFields := updateCustomFieldsFromAPI(resourceCustomFields, resource.CustomFields)
+			customFields := customfield.UpdateCustomFieldsFromAPI(resourceCustomFields, resource.CustomFields)
 
 			if err = d.Set("custom_field", customFields); err != nil {
 				return diag.FromErr(err)
@@ -225,7 +226,7 @@ func resourceNetboxTenancyTenantUpdate(ctx context.Context, d *schema.ResourceDa
 
 	if d.HasChange("custom_field") {
 		stateCustomFields, resourceCustomFields := d.GetChange("custom_field")
-		customFields := convertCustomFieldsFromTerraformToAPI(stateCustomFields.(*schema.Set).List(), resourceCustomFields.(*schema.Set).List())
+		customFields := customfield.ConvertCustomFieldsFromTerraformToAPI(stateCustomFields.(*schema.Set).List(), resourceCustomFields.(*schema.Set).List())
 		params.CustomFields = &customFields
 	}
