+++ netbox/resource_netbox_virtualization_vm.go
@@ -13,6 +13,7 @@ import (
 	netboxclient "github.com/smutel/go-netbox/v3/netbox/client"
 	"github.com/smutel/go-netbox/v3/netbox/client/virtualization"
 	"github.com/smutel/go-netbox/v3/netbox/models"
+	"github.com/smutel/terraform-provider-netbox/v4/netbox/internal/customfield"
 )
 
 func resourceNetboxVirtualizationVM() *schema.Resource {
@@ -45,7 +46,7 @@ func resourceNetboxVirtualizationVM() *schema.Resource {
 				Computed:    true,
 				Description: "The content type of this VM (virtualization module).",
 			},
-			"custom_field": &customFieldSchema,
+			"custom_field": &customfield.CustomFieldSchema,
 			"disk": {
 				Type:        schema.TypeInt,
 				Optional:    true,
@@ -158,7 +159,7 @@ func resourceNetboxVirtualizationVMCreate(ctx context.Context, d *schema.Resourc
 	clusterID := int64(d.Get("cluster_id").(int))
 	comments := d.Get("comments").(string)
 	resourceCustomFields := d.Get("custom_field").(*schema.Set).List()
-	customFields := convertCustomFieldsFromTerraformToAPI(nil, resourceCustomFields)
+	customFields := customfield.ConvertCustomFieldsFromTerraformToAPI(nil, resourceCustomFields)
 	disk := int64(d.Get("disk").(int))
 	localContextData := d.Get("local_context_data").(string)
 	memory := int64(d.Get("memory").(int))
@@ -260,7 +261,7 @@ func resourceNetboxVirtualizationVMRead(ctx context.Context, d *schema.ResourceD
 			}
 
 			resourceCustomFields := d.Get("custom_field").(*schema.Set).List()
-			customFields := updateCustomFieldsFromAPI(resourceCustomFields, resource.CustomFields)
+			customFields := customfield.UpdateCustomFieldsFromAPI(resourceCustomFields, resource.CustomFields)
 
 			if err = d.Set("custom_field", customFields); err != nil {
 				return diag.FromErr(err)
@@ -411,7 +412,7 @@ func resourceNetboxVirtualizationVMUpdate(ctx context.Context, d *schema.Resourc
 
 	if d.HasChange("custom_field") {
 		stateCustomFields, resourceCustomFields := d.GetChange("custom_field")
-		customFields := convertCustomFieldsFromTerraformToAPI(stateCustomFields.(*schema.Set).List(), resourceCustomFields.(*schema.Set).List())
+		customFields := customfield.ConvertCustomFieldsFromTerraformToAPI(stateCustomFields.(*schema.Set).List(), resourceCustomFields.(*schema.Set).List())
 		params.CustomFields = &customFields
 	}
